- Add new Backend method, counter(), which increments a named counter and   returns the new value. - Add gnc_book_get_counter() which implements the API and calls the   counter() method in the backend if it exists.  If not, it stores the   counters in the Book KVP Frame. - Modify all the Business objects to use the new counter API.  Counter is   gint64 so modify all the users to expect that.
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@6759 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/business/business-core/gncCustomerP.h
@@ -10,7 +10,7 @@
 #include "gncCustomer.h"
 
 gboolean gncCustomerRegister (void);
-gint gncCustomerNextID (GNCBook *book);
+gint64 gncCustomerNextID (GNCBook *book);
 void gncCustomerSetGUID (GncCustomer *customer, const GUID *guid);
 
 #endif /* GNC_CUSTOMERP_H_ */
