+++ src/business/business-core/gncJobP.h
@@ -10,7 +10,7 @@
 #include "gncJob.h"
 
 gboolean gncJobRegister (void);
-gint gncJobNextID (GNCBook *book);
+gint64 gncJobNextID (GNCBook *book);
 void gncJobSetGUID (GncJob *job, const GUID *guid);
 
 #endif /* GNC_JOBP_H_ */
